requirements.txt
discord.py==2.1.0
python-dotenv==0.21.0
openai==0.27.2
PyYAML==6.0
dacite==1.7.0
PyNaCl==1.5.0
Flask==2.2.2